refactor: fixes tash' comments
