+++ src/components/TeacherComponents/UnitList/UnitList.stories.tsx
@@ -11,7 +11,6 @@ import unitListingFixture, {
 
 const currentPageItems: UnitListProps = {
   ...unitListingFixture(),
-  selectedCategory: "biology",
   lessonCount: 10,
   currentPageItems: unitListingFixture().units.slice(0, 5),
   paginationProps: {
@@ -45,7 +44,6 @@ UnitList.args = currentPageItems;
 
 const currentPageItemsWithTiers: UnitListProps = {
   ...unitListingWithTiers(),
-  selectedCategory: "biology",
   lessonCount: 10,
   currentPageItems: unitListingWithTiers().units.slice(0, 5),
   paginationProps: {
